Moves currency formatter to stripe-ui-core.
