+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/AfterpayClearpayHeaderElement.kt
@@ -4,7 +4,7 @@ import android.content.res.Resources
 import androidx.annotation.RestrictTo
 import androidx.compose.ui.text.intl.Locale
 import com.stripe.android.ui.core.Amount
-import com.stripe.android.ui.core.CurrencyFormatter
+import com.stripe.android.uicore.format.CurrencyFormatter
 import com.stripe.android.ui.core.R
 import com.stripe.android.ui.core.forms.FormFieldEntry
 import kotlinx.coroutines.flow.Flow
